\name{MBCBExpressionData}
\alias{MBCBExpressionData} 
\alias{expressionSignal}
\alias{negativeControl}                                                                         

\title{MBCB - Bayesian Background Correction for Illumina Beadarray}

\description{
  The \code{MBCBExpressionData} dataset consists of two dataframes,
  \code{expressionSignal} and \code{negativeControl}. See the vignette
  for more detail regarding these data.
}

\usage{
  data(MBCBExpressionData)
}

\keyword{datasets}
